🎇feat: added ps2 mouse support

+++ kernel/src/Math.h
@@ -4,7 +4,7 @@ namespace Visionizer
 {
 	struct Point
 	{
-		unsigned int X;
-		unsigned int Y;	
+		long X;
+		long Y;	
 	};
 }
